Created a vertex displaced terrain mesh with basic lighting

+++ Core/src/Core/Utilities/CameraController3D.cpp
@@ -13,9 +13,9 @@ namespace KMG
 
 	void CameraController3D::OnUpdate(double dt)
 	{
-		float x = Input::GetKeyDown(Key::A) ? -1 : Input::GetKeyDown(Key::D) ? 1 : 0;
-		float y = Input::GetKeyDown(Key::C) ? -1 : Input::GetKeyDown(Key::Space) ? 1 : 0;
-		float z = Input::GetKeyDown(Key::S) ? -1 : Input::GetKeyDown(Key::W) ? 1 : 0;
+		int x = Input::GetKeyDown(Key::A) ? -1 : Input::GetKeyDown(Key::D) ? 1 : 0;
+		int y = Input::GetKeyDown(Key::C) ? -1 : Input::GetKeyDown(Key::Space) ? 1 : 0;
+		int z = Input::GetKeyDown(Key::S) ? -1 : Input::GetKeyDown(Key::W) ? 1 : 0;
 
 		glm::vec3 delta = m_MovementSpeed * glm::vec3(x, y, z) * static_cast<float>(dt);
